moved all mutations to individual files. Further testing of the API is needed.
