+++ server/schemas/mutations/index.js
@@ -1,0 +1,7 @@
+export { addUser, login } from "./user.js";
+export { addWine, updateWineDetails, removeWine, updateWineStorage } from "./wine.js";
+export { addProducer, updateProducer } from "./producer.js";
+export { addRegion, updateRegion } from "./region.js";
+export { addVariety, updateVariety } from "./variety.js";
+export { addComment, removeComment } from "./comment.js";
+export { addStorage } from "./storage.js";
